+++ src/state.rs
@@ -218,6 +218,7 @@ pub struct ItemInfo {
     pub preview_scroll: usize,
     pub content: Option<String>,
     pub permissions: Option<u32>,
+    pub is_dirty: bool,
 }
 
 #[derive(Default, Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
@@ -297,6 +298,7 @@ impl State {
                     dir_fg: color.dir_fg,
                     file_fg: color.file_fg,
                     symlink_fg: color.symlink_fg,
+                    dirty_fg: color.dirty_fg,
                 },
                 sort_by: session.sort_by,
                 show_hidden: session.show_hidden,
@@ -899,26 +901,16 @@ impl State {
         target_dir: &Option<PathBuf>,
         name_set: &mut BTreeSet<String>,
     ) -> Result<PathBuf, FxError> {
-        match target_dir {
-            None => {
-                let rename = rename_file(&item.file_name, name_set);
-                let to = &self.current_dir.join(&rename);
-                if std::fs::copy(&item.file_path, to).is_err() {
-                    return Err(FxError::PutItem(item.file_path.clone()));
-                }
-                name_set.insert(rename);
-                Ok(to.to_path_buf())
-            }
-            Some(path) => {
-                let rename = rename_file(&item.file_name, name_set);
-                let to = &path.join(&rename);
-                if std::fs::copy(&item.file_path, to).is_err() {
-                    return Err(FxError::PutItem(item.file_path.clone()));
-                }
-                name_set.insert(rename);
-                Ok(to.to_path_buf())
-            }
+        let rename = rename_file(&item.file_name, name_set);
+        let to = match target_dir {
+            None => self.current_dir.join(&rename),
+            Some(path) => path.join(&rename),
+        };
+        if std::fs::copy(&item.file_path, &to).is_err() {
+            return Err(FxError::PutItem(item.file_path.clone()));
         }
+        name_set.insert(rename);
+        Ok(to.to_path_buf())
     }
 
     /// Put single directory recursively to current or target directory.
@@ -1111,12 +1103,12 @@ impl State {
         let current_dir = self.current_dir.display().to_string();
         if current_dir.bytes().len() >= header_space {
             let current_dir = shorten_str_including_wide_char(&current_dir, header_space);
-            set_color(&TermColor::ForeGround(&Colorname::Cyan));
+            set_color_current_dir();
             print!(" {}", current_dir);
             reset_color();
             return;
         } else {
-            set_color(&TermColor::ForeGround(&Colorname::Cyan));
+            set_color_current_dir();
             print!(" {}", current_dir);
             reset_color();
             header_space -= current_dir.len();
@@ -1124,25 +1116,21 @@ impl State {
 
         // If without the write permission, print [RO].
         if self.is_ro && header_space > 5 {
-            set_color(&TermColor::ForeGround(&Colorname::Red));
+            set_color_read_only();
             print!(" [RO]");
             reset_color();
             header_space -= 5;
         }
 
-        //If .git directory exists, get the branch information and print it.
-        let git = self.current_dir.join(".git");
-        if git.exists() {
-            let head = git.join("HEAD");
-            if let Ok(head) = std::fs::read(head) {
-                let branch: Vec<u8> = head.into_iter().skip(16).collect();
-                if let Ok(branch) = std::str::from_utf8(&branch) {
+        //If git repository exists, get the branch information and print it.
+        if let Ok(repo) = git2::Repository::open(&self.current_dir) {
+            if let Ok(head) = repo.head() {
+                if let Some(branch) = head.shorthand() {
                     if branch.len() + 4 <= header_space {
                         print!(" on ",);
-                        set_color(&TermColor::ForeGround(&Colorname::LightMagenta));
+                        set_color_git_repo();
                         print!("{}", branch.trim().bold());
                         reset_color();
-                    } else {
                     }
                 }
             }
@@ -1160,11 +1148,15 @@ impl State {
             file_name
         };
         let time = format_time(&item.modified);
-        let color = match item.file_type {
+        let mut color = match item.file_type {
             FileType::Directory => &self.layout.colors.dir_fg,
             FileType::File => &self.layout.colors.file_fg,
             FileType::Symlink => &self.layout.colors.symlink_fg,
         };
+        if item.is_dirty {
+            color = &self.layout.colors.dirty_fg;
+        }
+
         if self.layout.terminal_column < PROPER_WIDTH {
             if item.selected {
                 set_color(&TermColor::ForeGround(color));
@@ -1227,9 +1219,44 @@ impl State {
         let mut dir_v = Vec::new();
         let mut file_v = Vec::new();
 
+        // If git repository exists, get information of changed/untracked files.
+        let mut dirty_paths = BTreeSet::new();
+        if let Ok(repo) = git2::Repository::discover(&self.current_dir) {
+            let mut opts = git2::DiffOptions::new();
+            // When detecting dirty files, includes untracked files.
+            opts.include_untracked(true);
+            if let Ok(diff) = repo.diff_index_to_workdir(None, Some(&mut opts)) {
+                // Current directory does not always point to the root (e.g. in the child dir),
+                // so uses repo.path() and pop() here.
+                let mut root = repo.path().to_path_buf();
+                root.pop();
+                diff.foreach(
+                    &mut |x, _| {
+                        if let Some(new_file) = x.new_file().path() {
+                            let dirty_path = root.join(new_file);
+                            // Changes color of ancestors to show a directory may contain
+                            // dirty files.
+                            for ancestor in dirty_path.ancestors() {
+                                dirty_paths.insert(ancestor.to_owned());
+                            }
+                        }
+                        true
+                    },
+                    None,
+                    None,
+                    None,
+                )
+                // Ignores error to continue the update_list process.
+                .unwrap_or(());
+            }
+        }
+
         for entry in fs::read_dir(&self.current_dir)? {
             let e = entry?;
-            let entry = read_item(e);
+            let mut entry = read_item(e);
+            if dirty_paths.contains(&entry.file_path) {
+                entry.is_dirty = true;
+            }
             match entry.file_type {
                 FileType::Directory => dir_v.push(entry),
                 FileType::File | FileType::Symlink => file_v.push(entry),
@@ -1312,33 +1339,21 @@ impl State {
     /// Highlight matched items.
     pub fn highlight_matches(&mut self, keyword: &str) {
         for item in self.list.iter_mut() {
-            if item.file_name.contains(keyword) {
-                item.matches = true;
-            } else {
-                item.matches = false;
-            }
+            item.matches = item.file_name.contains(keyword);
         }
     }
 
     /// Select items from the top to current position.
     pub fn select_from_top(&mut self, start_pos: usize) {
         for (i, item) in self.list.iter_mut().enumerate() {
-            if i <= start_pos {
-                item.selected = true;
-            } else {
-                item.selected = false;
-            }
+            item.selected = i <= start_pos;
         }
     }
 
     /// Select items from the current position to bottom.
     pub fn select_to_bottom(&mut self, start_pos: usize) {
         for (i, item) in self.list.iter_mut().enumerate() {
-            if i < start_pos {
-                item.selected = false;
-            } else {
-                item.selected = true;
-            }
+            item.selected = i >= start_pos;
         }
     }
 
@@ -1880,6 +1895,7 @@ fn read_item(entry: fs::DirEntry) -> ItemInfo {
                 preview_scroll: 0,
                 content: None,
                 permissions,
+                is_dirty: false,
             }
         }
         Err(_) => ItemInfo {
@@ -1897,6 +1913,7 @@ fn read_item(entry: fs::DirEntry) -> ItemInfo {
             preview_scroll: 0,
             content: None,
             permissions: None,
+            is_dirty: false,
         },
     }
 }